+++ mage/src/main/java/mil/nga/giat/mage/preferences/LocationPreferencesActivity.java
@@ -103,8 +103,6 @@ public class LocationPreferencesActivity extends AppCompatActivity {
                     } else {
                         application.stopLocationService();
                     }
-
-                    application.createNotification();
                 }
             });